Fix warnings when which are treated as errors with -developer-build
Task-number: QTBUG-91163
(cherry picked from commit 7185a63b3c0bcc797d683c5070a799b69ade019b)

+++ src/qml/jsapi/qjsmanagedvalue.cpp
@@ -909,7 +909,7 @@ QJSValue QJSManagedValue::property(quint32 arrayIndex) const
 
     if (QV4::String *string = d->as<QV4::String>()) {
         const QString qString = string->toQString();
-        if (arrayIndex < qString.size())
+        if (arrayIndex < quint32(qString.size()))
             return qString.sliced(arrayIndex, 1);
         return QJSValue();
     }
